Fixed field access via reflection for structs

+++ blitz.mod/blitz_debug.h
@@ -35,7 +35,7 @@ struct BBDebugDecl{
 	const char       *name,*type_tag;
 	union{
 		BBString*    const_value;
-		unsigned int field_offset;
+		size_t       field_offset;
 		void*        var_address;
 		size_t       struct_size;
 	};
